+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+# v0.3 (2024-05-08)
+- new spinners: newton
+- renamed spinner `spinner` to `slash`
+
 # v0.2 (2024-05-01)
 - switched to camelCase for spinner names
 - new spinners